プレイヤーの移動途中。 mymath追加
