+++ DirectXGame/MyMath.h
@@ -3,153 +3,33 @@
 #include "Vector3.h"
 #include "cmath"
 
-Matrix4x4 Multiply(const Matrix4x4& m1, const Matrix4x4& m2) {
-	Matrix4x4 result{};
-	for (int row = 0; row < 4; ++row) {
-		for (int column = 0; column < 4; ++column) {
-			for (int i = 0; i < 4; ++i) {
-				result.m[row][column] += m1.m[row][i] * m2.m[i][column];
-			}
-		}
-	}
-	return result;
-}
+// Vector3の足し算
+Vector3 operator+(const Vector3& v1, const Vector3& v2);
+Vector3& operator+=(Vector3& v1, const Vector3& v2);
+
+// Vector3の引き算
+Vector3 operator-(const Vector3& v1, const Vector3& v2);
+Vector3 operator-=(Vector3& v1, const Vector3& v2);
+
+// Vector3の掛け算(スカラー)
+Vector3 operator*(const Vector3& v, float s);
+Vector3& operator*=(Vector3& v, float s);
+
+// Vector3同士の掛け算
+Vector3 operator*(const Vector3& v1, const Vector3& v2);
+Vector3& operator*=(Vector3& v1, const Vector3& v2);
+
+Matrix4x4 Multiply(const Matrix4x4& m1, const Matrix4x4& m2);
 // 平行移動行列
-Matrix4x4 MakeTranslateMatrix(const Vector3& translate) {
-	Matrix4x4 translateMatrix;
-
-	translateMatrix.m[0][0] = 1;
-	translateMatrix.m[0][1] = 0;
-	translateMatrix.m[0][2] = 0;
-	translateMatrix.m[0][3] = 0;
-	translateMatrix.m[1][0] = 0;
-	translateMatrix.m[1][1] = 1;
-	translateMatrix.m[1][2] = 0;
-	translateMatrix.m[1][3] = 0;
-	translateMatrix.m[2][0] = 0;
-	translateMatrix.m[2][1] = 0;
-	translateMatrix.m[2][2] = 1;
-	translateMatrix.m[2][3] = 0;
-	translateMatrix.m[3][0] = translate.x;
-	translateMatrix.m[3][1] = translate.y;
-	translateMatrix.m[3][2] = translate.z;
-	translateMatrix.m[3][3] = 1;
-	return translateMatrix;
-}
+// 平行移動行列
+Matrix4x4 MakeTranslateMatrix(const Vector3& translate);
 // 拡大縮小行列
-Matrix4x4 MakeScaleMatrix(const Vector3& scale) {
-	Matrix4x4 scaleMatrix;
-	scaleMatrix.m[0][0] = scale.x;
-	scaleMatrix.m[0][1] = 0.0f;
-	scaleMatrix.m[0][2] = 0.0f;
-	scaleMatrix.m[0][3] = 0.0f;
-	scaleMatrix.m[1][0] = 0.0f;
-	scaleMatrix.m[1][1] = scale.y;
-	scaleMatrix.m[1][2] = 0.0f;
-	scaleMatrix.m[1][3] = 0.0f;
-	scaleMatrix.m[2][0] = 0.0f;
-	scaleMatrix.m[2][1] = 0.0f;
-	scaleMatrix.m[2][2] = scale.z;
-	scaleMatrix.m[2][3] = 0.0f;
-	scaleMatrix.m[3][0] = 0.0f;
-	scaleMatrix.m[3][1] = 0.0f;
-	scaleMatrix.m[3][2] = 0.0f;
-	scaleMatrix.m[3][3] = 1.0f;
-	return scaleMatrix;
-}
+Matrix4x4 MakeScaleMatrix(const Vector3& scale);
 // x軸回転行列
-Matrix4x4 MakeRotateXMatrix(float radian) {
-	Matrix4x4 result;
-	float c = cos(radian);
-	float s = sin(radian);
-
-	result.m[0][0] = 1.0f;
-	result.m[0][1] = 0.0f;
-	result.m[0][2] = 0.0f;
-	result.m[0][3] = 0.0f;
-
-	result.m[1][0] = 0.0f;
-	result.m[1][1] = c;
-	result.m[1][2] = s;
-	result.m[1][3] = 0.0f;
-
-	result.m[2][0] = 0.0f;
-	result.m[2][1] = -s;
-	result.m[2][2] = c;
-	result.m[2][3] = 0.0f;
-
-	result.m[3][0] = 0.0f;
-	result.m[3][1] = 0.0f;
-	result.m[3][2] = 0.0f;
-	result.m[3][3] = 1.0f;
-
-	return result;
-}
+Matrix4x4 MakeRotateXMatrix(float radian);
 // y軸回転行列
-Matrix4x4 MakeRotateYMatrix(float radian) {
-	Matrix4x4 result;
-	float c = cos(radian);
-	float s = sin(radian);
-
-	result.m[0][0] = c;
-	result.m[0][1] = 0.0f;
-	result.m[0][2] = -s;
-	result.m[0][3] = 0.0f;
-
-	result.m[1][0] = 0.0f;
-	result.m[1][1] = 1.0f;
-	result.m[1][2] = 0.0f;
-	result.m[1][3] = 0.0f;
-
-	result.m[2][0] = s;
-	result.m[2][1] = 0.0f;
-	result.m[2][2] = c;
-	result.m[2][3] = 0.0f;
-
-	result.m[3][0] = 0.0f;
-	result.m[3][1] = 0.0f;
-	result.m[3][2] = 0.0f;
-	result.m[3][3] = 1.0f;
-
-	return result;
-}
+Matrix4x4 MakeRotateYMatrix(float radian);
 // z軸回転行列
-Matrix4x4 MakeRotateZMatrix(float radian) {
-	Matrix4x4 result;
-	float c = cos(radian);
-	float s = sin(radian);
-
-	result.m[0][0] = c;
-	result.m[0][1] = s;
-	result.m[0][2] = 0.0f;
-	result.m[0][3] = 0.0f;
-
-	result.m[1][0] = -s;
-	result.m[1][1] = c;
-	result.m[1][2] = 0.0f;
-	result.m[1][3] = 0.0f;
-
-	result.m[2][0] = 0.0f;
-	result.m[2][1] = 0.0f;
-	result.m[2][2] = 1.0f;
-	result.m[2][3] = 0.0f;
-
-	result.m[3][0] = 0.0f;
-	result.m[3][1] = 0.0f;
-	result.m[3][2] = 0.0f;
-	result.m[3][3] = 1.0f;
-
-	return result;
-}
-// 3次元アフィン変換行列
-Matrix4x4 MakeAffineMatrix(const Vector3& scale, const Vector3& rotate, const Vector3& translate) {
-	Matrix4x4 scaleMatrix = MakeScaleMatrix(scale);
-	Matrix4x4 rotateXMatrix = MakeRotateXMatrix(rotate.x);
-	Matrix4x4 rotateYMatrix = MakeRotateYMatrix(rotate.y);
-	Matrix4x4 rotateZMatrix = MakeRotateZMatrix(rotate.z);
-	Matrix4x4 rotateMatrix = Multiply(rotateXMatrix, Multiply(rotateYMatrix, rotateZMatrix));
-	Matrix4x4 translateMatrix = MakeTranslateMatrix(translate);
-	Matrix4x4 affineMatrix = Multiply(Multiply(scaleMatrix, rotateMatrix), translateMatrix);
-
-	return affineMatrix;
-}
+Matrix4x4 MakeRotateZMatrix(float radian);
+// アフィン変換
+Matrix4x4 MakeAffineMatrix(const Vector3& scale, const Vector3& rotate, const Vector3& translate);
